+++ ext/nio4r/selector.c
@@ -25,7 +25,6 @@ static VALUE NIO_Selector_register(VALUE self, VALUE selectable, VALUE interest)
 static VALUE NIO_Selector_deregister(VALUE self, VALUE io);
 static VALUE NIO_Selector_is_registered(VALUE self, VALUE io);
 static VALUE NIO_Selector_select(int argc, VALUE *argv, VALUE self);
-static VALUE NIO_Selector_select_each(int argc, VALUE *argv, VALUE self);
 static VALUE NIO_Selector_wakeup(VALUE self);
 static VALUE NIO_Selector_close(VALUE self);
 static VALUE NIO_Selector_closed(VALUE self);
@@ -36,7 +35,6 @@ static VALUE NIO_Selector_unlock(VALUE lock);
 static VALUE NIO_Selector_register_synchronized(VALUE *args);
 static VALUE NIO_Selector_deregister_synchronized(VALUE *args);
 static VALUE NIO_Selector_select_synchronized(VALUE *args);
-static VALUE NIO_Selector_select_each_synchronized(VALUE *args);
 static int NIO_Selector_fill_ready_buffer(VALUE *args);
 static void NIO_Selector_timeout_callback(struct ev_loop *ev_loop, struct ev_timer *timer, int revents);
 static void NIO_Selector_wakeup_callback(struct ev_loop *ev_loop, struct ev_io *io, int revents);
@@ -59,7 +57,6 @@ void Init_NIO_Selector()
     rb_define_method(cNIO_Selector, "deregister", NIO_Selector_deregister, 1);
     rb_define_method(cNIO_Selector, "registered?", NIO_Selector_is_registered, 1);
     rb_define_method(cNIO_Selector, "select", NIO_Selector_select, -1);
-    rb_define_method(cNIO_Selector, "select_each", NIO_Selector_select_each, -1);
     rb_define_method(cNIO_Selector, "wakeup", NIO_Selector_wakeup, 0);
     rb_define_method(cNIO_Selector, "close", NIO_Selector_close, 0);
     rb_define_method(cNIO_Selector, "closed?", NIO_Selector_closed, 0);
@@ -271,28 +268,6 @@ static VALUE NIO_Selector_select(int argc, VALUE *argv, VALUE self)
     return NIO_Selector_synchronize(self, NIO_Selector_select_synchronized, args);
 }
 
-/* Select from all registered IO objects */
-static VALUE NIO_Selector_select_each(int argc, VALUE *argv, VALUE self)
-{
-    VALUE timeout, array;
-    VALUE args[2];
-
-    if(!rb_block_given_p()) {
-        rb_raise(rb_eArgError, "no block given");
-    }
-
-    rb_scan_args(argc, argv, "01", &timeout);
-
-    if(timeout != Qnil && NUM2DBL(timeout) < 0) {
-        rb_raise(rb_eArgError, "time interval must be positive");
-    }
-
-    args[0] = self;
-    args[1] = timeout;
-
-    return NIO_Selector_synchronize(self, NIO_Selector_select_each_synchronized, args);
-}
-
 /* Internal implementation of select with the selector lock held */
 static VALUE NIO_Selector_select_synchronized(VALUE *args)
 {
@@ -312,25 +287,6 @@ static VALUE NIO_Selector_select_synchronized(VALUE *args)
             /* new4 memcpys the ready buffer */
             return rb_ary_new4(ready, selector->ready_buffer);
         }
-    } else {
-        return Qnil;
-    }
-}
-
-/* Internal implementation of select with the selector lock held */
-static VALUE NIO_Selector_select_each_synchronized(VALUE *args)
-{
-    struct NIO_Selector *selector;
-    int i, ready = NIO_Selector_fill_ready_buffer(args);
-
-    Data_Get_Struct(args[0], struct NIO_Selector, selector);
-
-    if(ready > 0) {
-        for(i = 0; i < ready; i++) {
-            rb_yield(selector->ready_buffer[i]);
-        }
-
-        return INT2NUM(ready);
     } else {
         return Qnil;
     }