+++ spec/nio/selector_spec.rb
@@ -138,45 +138,6 @@ describe NIO::Selector do
     end
   end
 
-  context "select_each" do
-    it "iterates across ready selectables" do
-      readable1, writer = IO.pipe
-      writer << "ohai"
-
-      readable2, writer = IO.pipe
-      writer << "ohai"
-
-      unreadable, _ = IO.pipe
-
-      monitor1 = subject.register(readable1, :r)
-      monitor2 = subject.register(readable2, :r)
-      monitor3 = subject.register(unreadable, :r)
-
-      readables = []
-      subject.select_each { |monitor| readables << monitor }
-
-      readables.should include(monitor1)
-      readables.should include(monitor2)
-      readables.should_not include(monitor3)
-    end
-
-    it "allows new monitors to be registered in the select_each block" do
-      server = TCPServer.new("localhost", 10001)
-
-      monitor = subject.register(server, :r)
-      connector = TCPSocket.open("localhost", 10001)
-
-      block_fired = false
-      subject.select_each do |monitor|
-        block_fired = true
-        socket = server.accept
-        subject.register(socket, :r).should be_a NIO::Monitor
-      end
-
-      block_fired.should be_true
-    end
-  end
-
   it "closes" do
     subject.close
     subject.should be_closed